+++ src/main/java/com/bod/bod/user/controller/UserController.java
@@ -5,13 +5,9 @@ import com.bod.bod.global.dto.CommonResponseDto;
 import com.bod.bod.global.jwt.security.UserDetailsImpl;
 import com.bod.bod.user.dto.EditPasswordRequestDto;
 import com.bod.bod.user.dto.EditProfileRequestDto;
-import com.bod.bod.user.dto.LoginRequestDto;
 import com.bod.bod.user.dto.PointRankingResponseDto;
-import com.bod.bod.user.dto.SignUpRequestDto;
 import com.bod.bod.user.dto.UserResponseDto;
-import com.bod.bod.user.service.UserServiceImpl;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
+import com.bod.bod.user.service.UserService;
 import jakarta.validation.Valid;
 import java.util.List;
 import java.util.Map;
@@ -22,7 +18,6 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -39,49 +34,7 @@ import org.springframework.web.multipart.MultipartFile;
 @RequestMapping("/api")
 public class UserController {
 
-	private final UserServiceImpl userService;
-
-	@PostMapping("/signup")
-	public ResponseEntity<CommonResponseDto<Void>> signUp(
-		@RequestBody @Valid SignUpRequestDto signUpRequestDto
-	) {
-		userService.signUp(signUpRequestDto);
-		return ResponseEntity.ok().body(new CommonResponseDto<>
-			(HttpStatus.CREATED.value(), "회원가입이 완료되었습니다.", null));
-	}
-
-	@PostMapping("/login")
-	public ResponseEntity<CommonResponseDto<Void>> login(
-		@RequestBody @Valid LoginRequestDto loginRequestDto,
-		HttpServletResponse response
-	) {
-		userService.login(loginRequestDto, response);
-		return ResponseEntity.ok().body(new CommonResponseDto<>(
-			HttpStatus.OK.value(), "로그인이 완료되었습니다.", null));
-	}
-
-	@DeleteMapping("/logout")
-	public ResponseEntity<CommonResponseDto<Void>> logout(
-		@AuthenticationPrincipal UserDetailsImpl userDetails,
-		HttpServletRequest request,
-		HttpServletResponse response
-	) {
-		userService.logout(request, response, userDetails.getUser());
-		return ResponseEntity.ok().body(new CommonResponseDto<>(
-			HttpStatus.NO_CONTENT.value(), "로그아웃이 완료되었습니다.", null));
-	}
-
-	@DeleteMapping("/withdraw")
-	public ResponseEntity<CommonResponseDto<Void>> withdraw(
-		@RequestParam String username,
-		@RequestParam String password,
-		@AuthenticationPrincipal UserDetailsImpl userDetails,
-		HttpServletResponse response
-	) {
-		userService.withdraw(username, password, userDetails.getUser(), response);
-		return ResponseEntity.ok().body(new CommonResponseDto<>(
-			HttpStatus.OK.value(), "회원탈퇴가 완료되었습니다.", null));
-	}
+	private final UserService userService;
 
 	@PostMapping("/checkUserName")
 	public ResponseEntity<CommonResponseDto<Void>> validateUsernameRequest(
@@ -143,7 +96,7 @@ public class UserController {
 
 	@PutMapping("/users/profile/image")
 	public ResponseEntity<CommonResponseDto<UserResponseDto>> editProfileImage(
-		@RequestPart(value = "profileImage", required = false) MultipartFile profileImage,
+		@RequestPart(value = "profileImage") MultipartFile profileImage,
 		@AuthenticationPrincipal UserDetailsImpl userDetails
 	) {
 		UserResponseDto userResponseDto = userService.editProfileImage(profileImage, userDetails.getUser());
